+++ packages/plugin-packages/mashroom-security-provider-openid-connect/src/callback/mashroom-openid-connect-callback-route.ts
@@ -2,7 +2,7 @@ import openIDConnectClient from '../openid-connect-client';
 import createUser from '../create-user';
 import saveSession from '../save-session';
 import {OICD_REQUEST_DATA_SESSION_KEY_PREFIX, OICD_AUTH_DATA_SESSION_KEY, OICD_USER_SESSION_KEY} from '../constants';
-import type {OpenIDCallbackChecks,TokenSet} from 'openid-client';
+import type {OpenIDCallbackChecks, TokenSet} from 'openid-client';
 
 import type {Request, Response} from 'express';
 import type {MashroomLogger} from '@mashroom/mashroom/type-definitions';